requirements.txt
absl-py==2.1.0
accelerate==1.2.1
accelerator==2024.9.13
adam-atan2-pytorch==0.1.18
addict==2.4.0
aeiou==0.0.20
aiobotocore==2.16.1
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aioitertools==0.12.0
aiosignal==1.3.1
alembic==1.14.0
alias-free-torch==0.0.6
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
annoy==1.17.3
antlr4-python3-runtime==4.9.3
anyio==4.7.0
appdirs==1.4.4
argbind==0.3.9
asteroid-filterbanks==0.4.0
async-timeout==5.0.1
attrs==24.2.0
audioread==3.0.1
auraloss==0.4.0
banal==1.0.6
beartype==0.19.0
beautifulsoup4==4.12.3
bitsandbytes==0.45.0
bleach==6.2.0
bokeh==3.6.2
boto3==1.35.79
botocore==1.35.79
bottle==0.12.25
braceexpand==0.1.7
cached_path==1.6.5
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
clean-fid==0.1.35
click==8.1.7
clip-anytorch==2.6.0
colorcet==3.1.0
colorlog==6.9.0
configparser==7.1.0
contourpy==1.3.1
crcmod==1.7
cryptography==44.0.0
cycler==0.12.1
Cython==3.0.11
dataset==1.6.2
datasets==2.16.0
dctorch==0.1.2
decord==0.6.0
descript-audio-codec==1.0.0
descript-audiotools==0.7.2
diffusers==0.30.3
dill==0.3.7
Distance==0.1.3
docker-pycreds==0.4.0
docopt==0.6.2
docstring_parser==0.16
einops==0.8.0
einops-exts==0.0.4
einx==0.3.0
ema-pytorch==0.7.7
encodec==0.1.1
evaluate==0.4.3
fastapi==0.115.6
fastcore==1.7.28
fastdtw==0.3.4
ffmpy==0.5.0
filelock==3.13.1
fire==0.7.0
flatten-dict==0.4.2
fonttools==4.55.0
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2023.10.0
ftfy==6.3.1
future==1.0.0
g2p-en==2.1.0
gin-config==0.5.0
gitdb==4.0.11
GitPython==3.1.43
google==3.0.0
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-core==2.4.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
gradio==5.9.1
gradio_client==1.5.2
greenlet==3.1.1
grpcio==1.68.1
h11==0.14.0
h5py==3.12.1
holoviews==1.20.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.26.3
hydra-core==1.3.2
hyper-connections==0.1.8
HyperPyYAML==1.2.2
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
importlib-resources==5.12.0
inflect==7.4.0
inquirerpy==0.3.4
jaxtyping==0.2.36
jieba==0.42.1
Jinja2==3.1.3
jiwer==3.0.5
jmespath==0.10.0
joblib==1.4.2
jsonmerge==1.9.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
julius==0.2.7
k-diffusion==0.1.1
kiwisolver==1.4.7
kornia==0.7.4
kornia_rs==0.1.7
laion-clap==1.1.4
lazy_loader==0.4
librosa==0.9.2
lightning==2.4.0
lightning-utilities==0.11.9
linkify-it-py==2.0.3
llvmlite==0.43.0
local-attention==1.8.6
loguru==0.7.2
Mako==1.3.8
Markdown==3.7
markdown-it-py==3.0.0
markdown2==2.5.2
MarkupSafe==2.1.5
matplotlib==3.9.3
mdit-py-plugins==0.4.2
mdurl==0.1.2
modelscope==1.21.1
more-itertools==10.5.0
moviepy==2.1.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.15
networkx==3.2.1
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
omegaconf==2.3.0
opencv-python==4.10.0.84
optuna==4.1.0
orjson==3.10.12
oss2==2.19.1
packaging==24.1
pandas==2.0.2
panel==1.5.5
param==2.2.0
pathtools==0.1.2
pedalboard==0.7.4
pfzy==0.3.4
pillow==10.2.0
plotly==5.24.1
pooch==1.8.2
prefigure==0.0.9
primePy==1.3
proglog==0.1.10
progressbar==2.5
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.2
psutil==6.1.0
py==1.11.0
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyarrow==18.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.9.9
pydantic==2.10.4
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.18.0
pyloudnorm==0.1.1
pymcd==0.2.1
pynndescent==0.5.13
pyparsing==3.2.0
pypinyin==0.53.0
pysptk==1.0.1
pystoi==0.4.1
python-dotenv==1.0.1
python-multipart==0.0.20
python_speech_features==0.6
pytorch-lightning==2.1.0
pytorch-metric-learning==2.8.1
pytz==2024.2
pyviz_comms==3.0.3
PyWavelets==1.4.1
pyworld==0.3.4
PyYAML==6.0.2
randomname==0.2.1
RapidFuzz==3.11.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
resampy==0.4.3
retry==0.9.2
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
ruff==0.8.4
s3fs==2024.12.0
s3transfer==0.10.4
safehttpx==0.1.6
safetensors==0.4.5
scenedetect==0.6.5.2
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
semantic-version==2.10.0
semver==3.0.2
sentencepiece==0.1.99
sentry-sdk==2.19.0
setproctitle==1.3.4
shellingham==1.5.4
simplejson==3.19.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
SoundFile==0.10.2
soupsieve==2.6
soxr==0.5.0.post1
speechbrain==1.0.2
SQLAlchemy==1.4.54
starlette==0.41.3
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.12.12
tiktoken==0.8.0
tokenizers==0.21.0
tomli==2.2.1
tomlkit==0.13.2
torch==2.5.1+cu118
torch-audiomentations==0.11.1
torch-stoi==0.2.3
torch_pitch_shift==1.2.5
torchaudio==2.5.1+cu118
torchdiffeq==0.2.5
torchlibrosa==0.1.0
torchmetrics==0.11.4
torchsde==0.2.6
torchtyping==0.1.5
torchvision==0.20.1+cu118
tqdm==4.66.5
trampoline==0.1.2
transformers==4.47.1
transformers-stream-generator==0.0.5
triton==3.1.0
typeguard==4.4.1
typer==0.15.1
tzdata==2024.2
uc-micro-py==1.0.3
umap-learn==0.5.7
urllib3==2.2.3
uvicorn==0.34.0
v-diffusion-pytorch==0.0.2
vector-quantize-pytorch==1.9.14
vocos==0.1.0
volcengine==1.0.169
waitress==3.0.2
wandb==0.19.1
webdataset==0.2.48
webencodings==0.5.1
websockets==14.1
Werkzeug==3.1.3
wget==3.2
wrapt==1.17.0
x-transformers==1.43.5
xxhash==3.5.0
xyzservices==2024.9.0
yarl==1.18.3
zipp==3.20.2
